refactor(core): move interfaces used by Core to core module
