+++ CMakeLists.txt
@@ -270,7 +270,6 @@ set(${PROJECT_NAME}_SOURCES
   src/model/chathistory.h
   src/model/chathistory.cpp
   src/model/toxclientstandards.h
-  src/model/ibootstraplistgenerator.h
   src/model/notificationgenerator.h
   src/model/notificationgenerator.cpp
   src/net/bootstrapnodeupdater.cpp
