+++ core/CMakeLists.txt
@@ -22,12 +22,14 @@ set(SOURCE_FILES
     "include/core/dhtserver.h"
     "include/core/groupid.h"
     "include/core/groupinvite.h"
+    "include/core/ibootstraplistgenerator.h"
     "include/core/icorefriendmessagesender.h"
     "include/core/icoregroupmessagesender.h"
     "include/core/icoregroupquery.h"
     "include/core/icoreidhandler.h"
     "include/core/icoresettings.h"
     "include/core/itoxstatus.h"
+    "include/core/igroupsettings.h"
     "include/core/receiptnum.h"
     "include/core/toxcall.h"
     "include/core/toxencrypt.h"
