+++ core/include/core/igroupsettings.h
