+++ core/src/corefile.cpp
@@ -22,7 +22,6 @@
 #include "core/core.h"
 #include "core/toxfile.h"
 #include "toxstring.h"
-#include "src/model/status.h"
 #include "src/model/toxclientstandards.h"
 #include <QDebug>
 #include <QDir>
